Serialization and storage changes to Querier, Query, and Response.
 - Introduce serialization services that can flatten objects into a
variety of different formats.  Currently just reusing the Java
serialization code until the project is ready to adopt other
serializers.
 
 - Introduce storage services that can store objects and retrieve them.
Currently just reusing the local file system and HDFS.  The storage
service uses one of the serialization services to flatten the objects.
 
 - Remove code from Querier, Query, and Response that was serializing
and storing each of them; and update the callers to use the storage
services.