This is a WIP for [PIRK-63](https://issues.apache.org/jira/browse/PIRK-63) to open the door to other responders without having to modify the actual code of Pirk.  It's submitted for feedback only, please DO NOT MERGE.
It deprecates the "platform" CLI option in favor of the "launcher" option which is the name of a class implementing the `ResponderLauncher` interface which will invoke the run method via reflection.  This allows a developer of a different responder to merely place a jar on the classpath and specify the appropriate `ResponderLauncher` on the classpath.

The "platform" CLI option is still made available.  However, I removed the explicit dependencies in favor of using reflection.  This was done in anticipation other refactoring the build into submodules, though this does admittedly make the code more fragile.