Addressed comments from original commit per @tellison. -Now implemented as a Service Provider Interface. -Moved away from the -l (--launcher) CLI option in favor of -p (--platform), as we now have a platform method in the ResponderPlugin. This preserves backwards compatibility though it was not requested. -As mentioned on the mailing list the configuration is beyond the scope of this PR. -Minor testing was done standalone and spark work for sure.
Suggestions for a set good unit tests would be appreciated.  I think tests using something like hadoop-mini-cluster work be ideal.  I don't want to start that until after the build is broken up though as there will be multiple dependencies added to the tests.

Let me know if this is what we're looking for and I'll cleanup and test some more.