Correct the .travis.yml
The step that builds pirk is incorrectly in the install phase (which [Travis CI documentation](https://docs.travis-ci.com/user/customizing-the-build) says is to "install any dependencies required"). The script phase is to "run the build script" which is what was happening in the install phase. 

Right now the same tests run twice which increases build time for no benefit. To verify this, look at [this build log](https://s3.amazonaws.com/archive.travis-ci.org/jobs/174076031/log.txt) and search for `T E S T S`.